Added support for optionnal Encoding.
Before, the plugin generated every request in double: one URL encoded at
the minimum, the other one with a more "heavier" encoding.
As it can be not that useful, this option will permit to encode only if
asked to.
By default, it will only send one request for each test, with the
minimum encoding needed.